Added delay as an command option
For some devices (SHT30_I2CS), a pause is required between the command to take a reading and the actual reading back of the status. While this is better handled in the loop by checking for elapsed time and publishing a variable, that isn't possible during a function call -> return. Delay should be used sparingly, or not at all when it is possible to write your own code to handle this. HOWEVER, our goal with this library is to provide a generic library to allow you to easily communicate to ANY I2C device, and in that spirit delay will be supported.